refactor(useaccordion): move it to Accordion folder
