+++ src/components/organisms/Accordion/Accordion/Accordion.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import Accordion from '..';
+import { Accordion } from '..';
 
 describe('<Accordion />', () => {
   it('should render aria-label', () => {
