+++ src/components/organisms/Accordion/AccordionSection/AccordionSection.test.tsx
@@ -1,13 +1,13 @@
 import { axe } from 'jest-axe';
 import React from 'react';
 import { render } from '@testing-library/react';
-import Accordion from '..';
+import { Accordion } from '..';
 
 describe('<AccordionSection />', () => {
   it('renders the component with no a11y violations', async () => {
     const { container, getByTestId } = render(
       <Accordion aria-label="test-accordion">
-        <Accordion.Header id="one" index={0} textSize="medium">
+        <Accordion.Header id="one" index={0} textSize="base">
           Header
         </Accordion.Header>
         <Accordion.Section data-testid="section" id="one" index={0}>
