+++ src/components/organisms/Accordion/AccordionSection/AccordionSection.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 
-import { useAccordionContext } from '../context';
+import { useAccordionContext } from '../useAccordion';
 
 interface IAccordionSectionProps extends React.HTMLAttributes<HTMLDivElement> {
   id: string;
