--- src/components/organisms/Accordion/context/index.ts
@@ -1,2 +1,0 @@
-export { default as AccordionContext } from './AccordionContext';
-export { default as useAccordionContext } from './useAccordionContext';
