+++ src/components/organisms/Accordion/useAccordion/useAccordionContext.ts
@@ -1,13 +1,11 @@
 import { useContext } from 'react';
 
-import AccordionContext from './AccordionContext';
+import { AccordionContext } from './AccordionContext';
 
-const useAccordionContext = () => {
+export const useAccordionContext = () => {
   const context = useContext(AccordionContext);
   if (context === undefined) {
     throw new Error('useAccordionContext must be used within an Accordion');
   }
   return context;
 };
-
-export default useAccordionContext;
